refactor: QnABox에서 MultipleChoiceQuestion 분리
