+++ frontend/src/pages/ReviewWritingCardFromPage/components/QnABox/index.tsx
@@ -3,6 +3,8 @@ import { useMultipleChoice, useTextAnswer, useUpdateReviewerAnswer } from '@/hoo
 import useModals from '@/hooks/useModals';
 import { ReviewWritingCardQuestion } from '@/types';
 
+import MultipleChoiceQuestion from '../MultipleChoiceQuestion';
+
 import * as S from './style';
 
 const MODAL_KEY = {
@@ -17,32 +19,27 @@ interface QnABoxProps {
 
 const QnABox = ({ question }: QnABoxProps) => {
   const { updateAnswerMap, updateAnswerValidationMap } = useUpdateReviewerAnswer();
-  const { isOpen, openModal, closeModal } = useModals();
 
-  const handleModalOpen = (isOpen: boolean) => {
-    isOpen ? openModal(MODAL_KEY.confirm) : closeModal(MODAL_KEY.confirm);
-  };
+  /**
+   * 객관식 문항의 최소,최대 개수에 대한 안내 문구
+   */
+  const multipleLGuideline = (() => {
+    const { optionGroup } = question;
+    if (!optionGroup) return;
 
-  const { isOpenLimitGuide, multipleLGuideline, handleCheckboxChange, isSelectedCheckbox, unCheckTargetOption } =
-    useMultipleChoice({
-      question,
-      handleModalOpen,
-    });
+    const { minCount, maxCount } = optionGroup;
+
+    const isAllSelectAvailable = maxCount === optionGroup.options.length;
+    if (!maxCount || isAllSelectAvailable) return `(최소 ${minCount}개 이상)`;
+
+    return `(${minCount}개 ~ ${maxCount}개)`;
+  })();
 
   const { textAnswer, handleTextAnswerChange, TEXT_ANSWER_LENGTH } = useTextAnswer({
     question,
     updateAnswerMap,
     updateAnswerValidationMap,
   });
-
-  const handleModalCancelButtonClick = () => {
-    closeModal(MODAL_KEY.confirm);
-  };
-
-  const handleModalConfirmButtonClick = () => {
-    unCheckTargetOption();
-    closeModal(MODAL_KEY.confirm);
-  };
 
   return (
     <S.QnASection>
@@ -53,23 +50,7 @@ const QnABox = ({ question }: QnABoxProps) => {
       </S.QuestionTitle>
       {question.guideline && <S.QuestionGuideline>{question.guideline}</S.QuestionGuideline>}
       {/*객관식*/}
-      {question.questionType === 'CHECKBOX' && (
-        <>
-          {question.optionGroup?.options.map((option) => (
-            <CheckboxItem
-              key={option.optionId}
-              id={option.optionId.toString()}
-              isChecked={isSelectedCheckbox(option.optionId)}
-              disabled={false}
-              label={option.content}
-              handleChange={handleCheckboxChange}
-            />
-          ))}
-          <S.LimitGuideMessage>
-            {isOpenLimitGuide && <p>😅 최대 {question.optionGroup?.maxCount}개까지 선택가능해요.</p>}
-          </S.LimitGuideMessage>
-        </>
-      )}
+      {question.questionType === 'CHECKBOX' && <MultipleChoiceQuestion question={question} />}
 
       {/*서술형*/}
       {question.questionType === 'TEXT' && (
@@ -80,12 +61,6 @@ const QnABox = ({ question }: QnABoxProps) => {
           handleTextareaChange={handleTextAnswerChange}
           required={question.required}
         />
-      )}
-      {isOpen(MODAL_KEY.confirm) && (
-        <div>
-          <button onClick={handleModalConfirmButtonClick}>확인</button>
-          <button onClick={handleModalCancelButtonClick}>취소</button>
-        </div>
       )}
     </S.QnASection>
   );
